move node, building, inlines out of admin.py
