+++ src/meshapi/admin/admin.py
@@ -6,243 +6,11 @@ from nonrelated_inlines.admin import NonrelatedTabularInline
 
 from meshapi.models import Building, Device, Install, Link, Member, Node, Sector
 from meshapi.widgets import DeviceIPAddressWidget, PanoramaViewer
+from meshapi.admin.inlines import *
 
 admin.site.site_header = "MeshDB Admin"
 admin.site.site_title = "MeshDB Admin Portal"
 admin.site.index_title = "Welcome to MeshDB Admin Portal"
-
-
-# Inline with the typical rules we want + Formatting
-class BetterInline(admin.TabularInline):
-    extra = 0
-    can_delete = False
-    template = "admin/install_tabular.html"
-
-    def has_add_permission(self, request, obj) -> bool:
-        return False
-
-    class Media:
-        css = {
-            "all": ("admin/install_tabular.css",),
-        }
-
-
-class BetterNonrelatedInline(NonrelatedTabularInline):
-    extra = 0
-    can_delete = False
-    template = "admin/install_tabular.html"
-
-    def has_add_permission(self, request, obj) -> bool:
-        return False
-
-    class Media:
-        css = {
-            "all": ("admin/install_tabular.css",),
-        }
-
-
-class NonrelatedBuildingInline(BetterNonrelatedInline):
-    model = Building
-    fields = ["primary_node", "bin", "street_address", "city", "zip_code"]
-    readonly_fields = fields
-
-    add_button = True
-
-    # Hack to get the NN
-    network_number = None
-
-    def get_form_queryset(self, obj):
-        self.network_number = obj.pk
-        return self.model.objects.filter(nodes=obj)
-
-    def save_new_instance(self, parent, instance):
-        pass
-
-
-class BuildingMembershipInline(admin.TabularInline):
-    model = Building.nodes.through
-    extra = 0
-    autocomplete_fields = ["building_id"]
-    classes = ["collapse"]
-    verbose_name = "Building"
-    verbose_name_plural = "Edit Related Buildings"
-
-
-# This controls the list of installs reverse FK'd to Buildings and Members
-class InstallInline(BetterInline):
-    model = Install
-    fields = ["status", "node", "member", "building", "unit"]
-    readonly_fields = fields
-
-
-class DeviceInline(BetterInline):
-    model = Device
-    fields = ["status", "type", "model"]
-    readonly_fields = fields
-
-    def get_queryset(self, request):
-        # Get the base queryset
-        queryset = super().get_queryset(request)
-        # Filter out sectors
-        queryset = queryset.exclude(sector__isnull=False)
-        return queryset
-
-
-class NodeLinkInline(BetterNonrelatedInline):
-    model = Link
-    fields = ["status", "type", "from_device", "to_device"]
-    readonly_fields = fields
-
-    def get_form_queryset(self, obj):
-        from_device_q = Q(from_device__in=obj.devices.all())
-        to_device_q = Q(to_device__in=obj.devices.all())
-        all_links = from_device_q | to_device_q
-        return self.model.objects.filter(all_links)
-
-    def save_new_instance(self, parent, instance):
-        pass
-
-
-class DeviceLinkInline(BetterNonrelatedInline):
-    model = Link
-    fields = ["status", "type", "from_device", "to_device"]
-    readonly_fields = fields
-
-    def get_form_queryset(self, obj):
-        from_device_q = Q(from_device=obj)
-        to_device_q = Q(to_device=obj)
-        all_links = from_device_q | to_device_q
-        return self.model.objects.filter(all_links)
-
-    def save_new_instance(self, parent, instance):
-        pass
-
-
-class SectorInline(BetterInline):
-    model = Sector
-    fields = ["status", "type", "model"]
-    readonly_fields = fields
-
-
-class BoroughFilter(admin.SimpleListFilter):
-    title = "Borough"
-    parameter_name = "borough"
-
-    def lookups(self, request, model_admin):
-        return (
-            ("bronx", ("The Bronx")),
-            ("manhattan", ("Manhattan")),
-            ("brooklyn", ("Brooklyn")),
-            ("queens", ("Queens")),
-            ("staten_island", ("Staten Island")),
-        )
-
-    def queryset(self, request, queryset):
-        if self.value() == "bronx":
-            return queryset.filter(city="Bronx")
-        elif self.value() == "manhattan":
-            return queryset.filter(city="New York")
-        elif self.value() == "brooklyn":
-            return queryset.filter(city="Brooklyn")
-        elif self.value() == "queens":
-            return queryset.filter(city="Queens")
-        elif self.value() == "staten_island":
-            return queryset.filter(city="Staten Island")
-        return queryset
-
-
-class BuildingAdminForm(forms.ModelForm):
-    class Meta:
-        model = Building
-        fields = "__all__"
-        widgets = {
-            "panoramas": PanoramaViewer(schema={"type": "array", "items": {"type": "string"}}),
-            "bin": forms.NumberInput(attrs={"style": "width:21ch"}),
-        }
-
-
-@admin.register(Building)
-class BuildingAdmin(admin.ModelAdmin):
-    form = BuildingAdminForm
-    list_display = ["__str__", "street_address", "primary_node"]
-    search_fields = [
-        # Sometimes they have an actual name
-        "nodes__name__icontains",
-        # Address info
-        "street_address__icontains",
-        "city__icontains",
-        "state__icontains",
-        "zip_code__iexact",
-        "bin__iexact",
-        # Search by NN
-        "nodes__network_number__iexact",
-        "installs__install_number__iexact",
-        # Search by Member info
-        "installs__member__name__icontains",
-        "installs__member__primary_email_address__icontains",
-        "installs__member__phone_number__iexact",
-        "installs__member__slack_handle__iexact",
-        # Notes
-        "notes__icontains",
-    ]
-    list_filter = [
-        BoroughFilter,
-        ("primary_node", admin.EmptyFieldListFilter),
-    ]
-    fieldsets = [
-        (
-            "Address Information",
-            {
-                "fields": [
-                    "street_address",
-                    "city",
-                    "state",
-                    "zip_code",
-                ]
-            },
-        ),
-        (
-            "NYC Information",
-            {
-                "fields": [
-                    "bin",
-                    "latitude",
-                    "longitude",
-                    "altitude",
-                ]
-            },
-        ),
-        (
-            "Misc",
-            {
-                "fields": [
-                    "notes",
-                    "panoramas",
-                ]
-            },
-        ),
-        (
-            "Nodes",
-            {
-                "fields": [
-                    "primary_node",
-                    "nodes",
-                ]
-            },
-        ),
-    ]
-    inlines = [InstallInline]
-    filter_horizontal = ("nodes",)
-
-    # This is probably a bad idea because you'll have to load a million panos
-    # and OOM your computer
-    # Need to find a way to "thumbnail-ize" them on the server side, probably.
-    @mark_safe
-    def thumb(self, obj):
-        return f"<img src='{obj.get_thumb()}' width='50' height='50' />"
-
-    thumb.__name__ = "Thumbnail"
-
 
 class MemberAdminForm(forms.ModelForm):
     class Meta:
@@ -457,76 +225,6 @@ class LinkAdmin(admin.ModelAdmin):
     list_filter = ["status", "type"]
 
     autocomplete_fields = ["from_device", "to_device"]
-
-
-class NodeAdminForm(forms.ModelForm):
-    class Meta:
-        model = Node
-        fields = "__all__"
-
-
-@admin.register(Node)
-class NodeAdmin(admin.ModelAdmin):
-    form = NodeAdminForm
-    search_fields = [
-        "network_number__iexact",
-        "name__icontains",
-        "buildings__street_address__icontains",
-        "notes__icontains",
-    ]
-    list_filter = ["status", ("name", admin.EmptyFieldListFilter), "install_date", "abandon_date"]
-    list_display = ["__network_number__", "name", "status", "address", "install_date"]
-    fieldsets = [
-        (
-            "Details",
-            {
-                "fields": [
-                    "status",
-                    "type",
-                    "name",
-                ]
-            },
-        ),
-        (
-            "Location",
-            {
-                "fields": [
-                    "latitude",
-                    "longitude",
-                    "altitude",
-                ]
-            },
-        ),
-        (
-            "Dates",
-            {
-                "fields": [
-                    "install_date",
-                    "abandon_date",
-                ]
-            },
-        ),
-        (
-            "Misc",
-            {
-                "fields": [
-                    "notes",
-                ]
-            },
-        ),
-    ]
-    inlines = [
-        InstallInline,
-        NonrelatedBuildingInline,
-        BuildingMembershipInline,
-        DeviceInline,
-        SectorInline,
-        NodeLinkInline,
-    ]
-
-    def address(self, obj):
-        return obj.buildings.first()
-
 
 device_fieldsets = [
     (
